% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FerariIndex_Difference.R
\docType{data}
\name{FerariIndex_Difference}
\alias{FerariIndex_Difference}
\title{FerariIndex_Difference of vine dataset}
\format{
A data frame with 32 observations and 1 variable.

\describe{
  \item{fi_dif}{numeric. Ferari index differences between July 24th, 2014 at 09:00 am 
  and August 01, 2014 at 09:00 am.}
}
}
\source{
These data were acquired during the Innovine project, funded by the Seventh
Framework Programme of the European Community (FP7/2007-2013), under Grant Agreement
No. FP7-311775.
}
\usage{
FerariIndex_Difference
}
\description{
Data were collected during an experiment conducted on a vineyard
of the INRAE/Institut Agro campus at Montpellier in 2014 (Syrah vines). The objective
of the experiment was to study the influence of the micro-climate (temperature and
irradiance) at the grape level on the anthocyanin contents of the berries indicated by
the Ferari index. This dataset contains Ferari index differences between August 01,
2014 at 09:00 am and July 24th, 2014 at 09:00 am. The
individuals are in rows. The individuals' names (Indiv1,...,Indiv32) are used to name the rows. 
The same individuals are also
present in the irradiance and temperaure datasets.
}
\keyword{datasets}
